feat: combine and refactor PO code
